data migration AddColumnToSurvey-
Got survey results linked to user account, user2. Had trouble getting
workouts and days to be matched with an account. There is some commented
out code in models and controllers that I was trying that failed. Also
some hidden fields I deleted from their corresponding views. Added
'Profile' to menu that links you to a show page, users/(current user).
On there is a link to a user's survey results.